NPF: Add matching against IP header protocol field
Jira: VRVDR-53280

This provides a restricted match capability akin to that which
can be supported in limited hardware matching facilities.  Hence
we can have the software ACL behave in the same fashion as the
hardware ACL on such platforms.

For IPv4 this is the IP header protocol field, for IPv6 the IP
header 'next header' field. The pre-existing match capability
already follows extension chains, and has been renamed to
proto-final.