fal_plugin: add object model for MPLS
Add support in the object model for expressing imposing labels for
IPv4 and IPv6 routes through the label stack attribute on next-hop
objects, which can express a list of labels.

Add support in the object model for MPLS routes using a new MPLS route
object whose key is expressed in a similar way to that for IP routes
in a structure, and which doesn't return a fal_object_t. The reason
for not returning a fal_object_t is for reasons of scale where it
isn't convenient for the FAL plugin to allocate an object for each
MPLS route and it may not have any other suitable ID that it can
returned instead. The attributes supported on MPLS routes follow that
for IP routes.

VRVDR-52435