mpls: signal FAL create/update/delete of MPLS routes
Add support for signalling the create, update and delete of MPLS
routes through the FAL. Store the last programming state to avoid
issuing an update or delete for a route that wasn't created, which
would otherwise violate the FAL API, and also allow that state to be
displayed to the user.

Add support for backwalking from next-hop-lists to MPLS routes upon
interfaces becoming L3 enabled/disabled so that the MPLS routes are
suitably updated to either refer to a FAL router-interface that has
just been created, or to stop referring to a FAL router-interface that
will be deleted.

VRVDR-52436