fal: signal use of next hop groups
It is beneficial to signal whether a next-hop-group will be used for
IP routing or label-switching since it has an impact on what to
program, at least whether to program a pop, but perhaps also whether
to program a swap vs. push.

This could be determined in the FAL plugin upon linkage from the
MPLS/IP route to next-hop-group, but would be onerous in terms on
state and code to do so.

So provide a hint that signals in advance the usage of the
next-hop-group and ensure it is only used in that context.

VRVDR-52436