ipsec-ut: Add a test to measure time to setup/teardown 500 tunnels
This test is being marked DONT_RUN in order to prevent it from
running automatically on the test execution server. The test is a scale
test and the execution time can vary depending on the load on the
execution server. If the execution server is too heavily loaded, the
test can take a long time to complete and result in spurious failures.
Marking it as a DONT_RUN test prevents such failures while keeping the
test available for development use.

VRVDR-48526