fal_plugin: add FAL object for VRFs
Add a FAL object for VRFs to signal creation and deletion explicitly
for any FAL plugins that require this, as well as providing an object
that can be referred to from the next-hop object to signal that
packets using the next-hop should perform a lookup in a VRF, rather
than being forwarded directly out of an interface towards a neighbour.

Make use of the new VRF object in the router-interface object,
deprecating the existing integer parameter. However, the new VRF
object isn't used for IP route objects at this time as this would
require a backwards incompatible change.

VRVDR-52435