crypto: Increase force commit count to 2000
The current value at which an NPF commit is forced is too low.
The overhead of each NPF rebuild is very high and has a huge
impact on rebuild time for large numbers of tunnels. Increasing
the batch size to approximately the tunnel setup rate being
targeted results in a significant reduction in rebuild time.
The details are documented in the LLD.

VRVDR-50487