protobuf: add definition for route updates from the RIB
Add a protobuf definition for route updates from the RIB, including
route, paths and labels. A unified route format is used for both MPLS
and IP routes to enable maximum code re-use, primarily in the RIB, but
hopefully in the longer term in the dataplane too.

Unlike netlink, no special distinction is made between a single path
case and multipath case, which simplifies the code for both encoding
and decoding. Blackhole/unreachable/local become path types to reflect
that both the dataplane and RIB deal with the forwarding on a
path-by-path basis, and at least on the dataplane side simplifies
decoding, as well as to allow flexibility for potential future use
cases, even if in the current use cases having a mixture of
blackhole/unreachable/local and unicast paths doesn't make sense.

The scope field is present due to certain cases where routes are
distributed directly from kernel netlink broadcasts instead of being
installed in the RIB and having the best-path selection algorithm run
over them. This means that there can be two routes present with the
same table & prefix and the scope is used to distinguish between this.

The routing_protocol field is not used by the dataplane, other than
for display purposes, but it also helps in the generation of the
netlink format of the route by the RIB and in the common case won't
take up any space in the message.

The preferred_source path field is not used by the dataplane, but it
helps in the efficient generation of the netlink format of the route
by the RIB, and it could be of use in the future. In the BGP scale
case it wouldn't be present and thus takes up no extra space in the
message and so doesn't adversely impact convergence.

Having a top-level RibUpdate message with a Route as a field allows
for future extensibility for other object updates (e.g. EVPN L2RIB
updates) to be received over the same channel without breaking
compatibility (it is allowed to convert an optional field into a oneof
with that field as a member).

VRVDR-50720