ip_rt_protobuf: add support for decoding protobuf route updates
Add support for receiving updates from a RIB in protobuf format, in a
backwards compatible way. The netlink handling is retained for three
purposes:
1. Multicast updates which aren't represented in the protobuf format.
2. For RIBs not yet able to generate our protobuf format of messages.
3. Any future remote dataplane functionality where there may not be a
   RIB running locally, which would otherwise be required for
   boostrapping connectivity.

Since this path is somewhat performance sensitive during convergence
at scale, the protobuf_get_ipaddr helper isn't used for handling of
nexthop addresses as that would potentially involve another data copy
(if not optimised away by compiler).

The incomplete route handling is adapted to be able to replay protobuf
messages as well as netlink messages, depending on the form they
arrive in.

VRVDR-50720