ip_rt_protobuf: add support for installing backup paths
Add a new path field "backup" which signals that a path is a PIC Edge
backup, i.e. not a 1:1 backup but that the backup path will only be
used when all primary paths have gone down.

Add support for installing backup paths, as signalled by the route
update protobuf message, and support for displaying such paths.

VRVDR-50720