fal: add FAL next-hop attributes for PIC Edge
Add support for attributes for PIC Edge to allow the hardware to
pre-program backup paths and perform a fast re-route in the case of
one or more primary path failures.

The usability attribute needs to be at a nexthop level in order to get
the required semantics that if one primary nexthop fails a cutover is
done to the remaining primary nexthop, and if that fails then the
backup nexthop(s) is/are used.

Closes: VRVDR-50739