ut: add a new test for qos bursts
Add a new test that sends 15 seconds of traffic at 50Mbits/s.
Apply a qos policy at 49.5Mbits/s and verify that all the packets
are forwarded, and that qos drops none.

Use the packet sending APIs that do no verification of the packet
contents for speed.

As this test is dependent on the cpu/load of the machine where it
is being run start it as a test that is marked as DONT_RUN. It is
added here though so that it can be used as a basis for future
qos testing/debugging as required.

VRVDR-50306