fal: add memory management functions
Add memory management functions to allow deferred freeing of allocated
memory for thread safety purposes. Having these wrappers avoids the
FAL plugin needing to be aware of the application (dataplane)
implementation of how this is done, such as a particular flavour of
URCU, and allows the FAL plugins to be insulated from any change to a
different mechanism with similar functionality.

The intention is that any objects that are used from functions invoked
other than on the master thread (e.g. mostly get/dump functions) are
allocated/freed using these functions, which typically will be just
the FAL objects themselves, but could include other memory.

To avoid the trap of freeing hardware resources (which are very
limited) from an RCU callback, no utility function is provided
currently to do this.

There will be a small amount of overhead in the memory allocation as a
result of using these wrappers compared to the bare libc calls before,
but this is the least worst option.

VRVDR-50764