alg: ensure parent session active before linking child
With the ALGs the child session locates the parent session by looking
up the tuple databases for a match. The child should not use the parent
session unless it has been fully set up (and so would also have the base
session structure associated), and this is indicated by the SE_ACTIVE
flag being set. The parent not being fully set-up can happen if a packet
relating to the parent is being processed on one core while another
packet relating to the child is being set-up at the same time on a
different core.

The code was updated to replace calls to apt_tuple_get_session() when
used to associate a child ALG session with a parent with new function
apt_tuple_get_active_session(), which ensures the parent session is
active or returns NULL. The code in the ALGs was also updated to check
for NULL being returned, and fail if it is, which indicates the child
session could not be created.

VRVDR-50613