vlan_modify: don't create filter chain unless there's an action we handle
Don't lookup and create a filter chain until after we've looking at
the action in the TC filter chain and have decided that it's a VLAN
modify action.

This avoids creating a filter chain and enabling the VLAN modify
feature when an action that we're not interested in, such as
Act_mirred, is applied which is always the case for switch vifs.

Fixes: VRVDR-50709