vlan_modify: remove some noisy log messages
Remove a duplicate log message emitted by vlan_mod_flt_lookup_filter,
which is the same as the one emitted in the calling function (minus
determining whether it's a new filter or an update).

Bump messages for mirred and unknown actions down to debug level,
since they are expected and we have no cause to need to handle them
for the purposes of VLAN modification.

Fixes: VRVDR-50711