route: make the gateway in a next hop a union of v4/v6
Start the transition of the next_hop to support both v4 and v6 so that
we don't have the same code repeated for both address families.

Change the gateway to be a union of v4 and v6 address.  This changes the
the size of the struct from 32 to 40 (there were 4 bytes of trailing
padding).

VRVDR-50724