route: change order of fields in struct next_hop_u
Change the order of the fields in the next_hop_u to remove holes in the
structure, while making sure that the fields that are used in the fast
path are in the first cache line.

Before these changes (and the changes to make the next_hop v4/v6) the
size of this structure was 128 bytes, with 12 bytes of holes.  After
this change it is still 128 bytes but now with no holes (although there
are 2 bytes of padding).

VRVDR-50724