route6: make the v6 route code use the common 'struct next_hop'
Make the v6 code use the struct next_hop which now has a union of
v4/v6 addresses.

To avoid changing the public APIs leave them using the old v6
specific version of the structure in the prototypes, but define
the v6 version to be the same as the general version.

VRVDR-50724