nh_common: remove struct next_hop_v6_u, use the v4/v6 version
Now that the struct_next_hop_u can be used for v4 and v6 NHs
there is no need to have a v6 specific version of the structure.
Change the v6 code to use the common structures.

The difference were in field layout, the common one has a proto
field where the v6 one does not and the v6 one had a larger
number of siblings allowed. However the number the common one
supports is enough for both v4 and v6.

VRVDR-50724