route: don't use global nh_tbl var from nexthop_new
nexthop_new was using the global nh_tbl (v4) structure when
creating a new nexthop. Add a family parameter to the func
and use that to retieve the correct table via a new api to
get the nh_table for a given family.

This is in preparation for a single func to create a new nexthop
for either v4 or v6.

VRVDR-50724