nh_common: add code to allow registration per AF
As we move towards a common infra for NHs there will be things that are
still different between address families. Add apis to allow each family
to register functions to deal with the differences.

Start by the addition of 3 functions to do hashing and 1 function to
get the nh_table for the address family. These are the set initially
needed to move the nexthop_new func to being fully common.

VRVDR-50724