ut: mark dp_test_crypto_perf_scale tests as DONT_RUN
The crypto perf scale tests are useful to be able to compare runs
on the same machine, but they can fail on slow machines or on
heavily loaded machines. As we need the tests that are part of the
build to always pass irrespective of the machine speed, don't run
these ones.

VRVDR-50753