nh_common: make the nexthop_lookup function common
Move the nh_lookup from route.c into nh_common. Provide a public API
for it so that both v4 and v6 can use it. Modify the code so that
instead of using the hard coded hash details it accesses the hash
details via the new nh_common apis.

VRVDR-50724