nh_common: add common funcs to get/set ifp from nh
There was a v4 and a v6 specific function to get and set the ifp
in a next_hop, and they were very similar. The differences were in
a log message and one had an extra rcu_assign_pointer.

Change to using a common function that is not AF specific. As the
public api contains functions for getting the ifp leave them there
and mark them as deprecated. They become wrappers around the new
func.

VRVDR-50724