fal: use a common version of next_hop_to_attr_list
There were 2 version of this func, one for v4 and one for v6 but
the code was very similar. The differences were in setting the
family in the address and copying the address.

Move to a single version of the function that takes the address
family, and uses this when setting the family. It also copies
the full address structure as this will make sure that the copy
works for both the v4 and v6 addresses.

VRVDR-50724