fal: use a common version of fal_ip_new_next_hops
There were 2 version of this func, one for v4 and one for v6 but
the code was very similar. The differences were that the family
being used wa hard coded.

Move to a single version that gets given the family parameter as
a parameter, and change the v6 and v4 route code to use it.

VRVDR-50724