nh_common: use a common version of nexthop_reuse
There were 2 version of this func, one for v4 and one for v6 but
the code was very similar. The differences were that one version
had a local var to store the index to use later in the debug, and
in the debug itself.

Move to a single version that gets given the family  as
a parameter, and change the v6 and v4 route code to use it.

VRVDR-50724