nh_common: use a common version of nexthop_hash_insert
There were 2 version of this func, one for v4 and one for v6 but
the code was very similar. The differences were in the hash table
and hash functiosn used

Move to a single version that gets given the family as
a parameter, and use that param to get the relevant hash state to
use.

VRVDR-50724