nh_common: use a common version of nexthop_new
There were 2 version of this func, one for v4 and one for v6 but
the code was very similar with the main difference being in extra
debug in v6.

Move to a single version that gets given the family as
a parameter, and keep the extra debug but slightly modified to
be af independent.

VRVDR-50724