nh_common: use a common version of nexthop_put
There were 2 version of this func, one for v4 and one for v6 but
the code was very similar. The differences were family specific
debug and accessing the family specific tables/hash directly.

Move to a single version that uses the family to retrieve the nh
table and the hash. Modify the debug to print the correct family

VRVDR-50724