nh_common: use a common version of nexthop_hash_del_add
There were 2 version of this func, one for v4 and one for v6 but
the code was the same apart from the way the hash table was accessed
and the address family being hard coded.

Move to a single version that is used by both v4 and v6 and get the
correct hash table by using the passed in family to retrieve it, and
pass the family to the called functions.

VRVDR-50724