nh_common: use a common version of nh_set_neigh_present
There were 2 version of this func, one for v4 and one for v6 but
the code was the same other than for the accessing the af specific
nh table.

Move to a single version that is used by both v4 and v6 and accesses
the correct nh table based on the family passed in.

VRVDR-50724