nat: add support in apm for multiple port maps based on protocol
Add support for multiple port maps (for 'tcp, 'udp, and 'other'
protocols), though currently only use the first one ('tcp') so that the
APIs to other code is not changed yet.

The unit tests needed updating as they do the 'npf fw dump-portmap'
vplsh command, and this has changed to give ports used and mapping_count
for each protocol.

VRVDR-50718