nat: add passing in IP proto to apm port request fns
Update the functions npf_apm_get_map(), npf_apm_put_map() and
npf_apm_get_allocated(), which alloc, release, and query ports to also
take an IP protocol as a parametr. They map the IP proto to a nat-proto
('tcp', 'udp', and 'other') to access the appropriate pool of ports.

Update functions which allocate ports in ALG and NAT to pass in the IP
protocol to the apm functions.  Typically the IP protocol comes from
with the npf cache, or an associated session.

Unit tests which query the APM using 'npf fw dump-portmap' are updated
to says which protocol to query the port mappings. Also a new unit
test has been added.

VRVDR-50718