nh_common: use a common version of nh_get_flags
This function was taking a union of the v4/v6 nexthop, but these
are the same thing. Change the signature to take a next_hop
and make it return the flags.

Change the callers to pass in the v4 version of the union.  This will
still work for the v6 case too, and will be modified further in
commits

VRVDR-50724