mpls: use dp_nh_get_ifp instead of nh_get_if
The former takes a struct next_hop, but the latter takes a union of
next hops (v4 and v6) and a type.  However both fields in the union
are now of the same type so it does not need to check which one it
is.  Pass in the v4 version for now, which will also work for the
v6 case. This will be modified further in later commits

This change means that nh_fwd_mpls no longer uses the the
next_hop_type passed in, so remove the parameter to keep it compiling.

VRVDR-50724