nat: update per-rule nat statistics to be per-protocol
The statistics stored against each rule is updated so it is tracked
per-protocol. Note that returning the per-protocol statistics using JSON
is not in this commit, but will be in the next one.

For DNAT it also totals the "used" value per-protocol. It no longer
limits them to the "total" value, making the code consistent with SNAT.

VRVDR-50718