ecmp: change ecmp_mpls_create to return a struct next_hop ptr
Make the function return a pointer to a next_hop rather than the
union of next hops for v4/v6.  The callers are changed, and when
they have the union they pass the v4 field from it, but this will
work for v6 too as they share the same type and storage.

VRVDR-50724