nh: move the final funcs from nh.c and delete the file
Move the final 2 funcs, the ones to get the ip addr from a
next_hop into the nh_common.c and delete nh.c

These 2 funcs are only kept for backwards compatibility with
the plugins. As the next_hop does not contain info to say
which address family it is using there is not currently a
common API to get the address.

VRVDR-50724