nat64: set full range for overload start/stop ports
For overload ports which are configured as a single address or a
group the start/stop ports are left as 0 rather than setting them to
1 and 65535 (which it does for an address range). This results in
npf_apm_create() choosing the defaults of 1 and 65535, however the
total amount of ports available is not calculated correctly by calls to
npf_natpolicy_get_map_range() due to them being 0.

This change moves the setting of start/stop ports from the "srange"
processing to just before the call to npf_create_natpolicy() in
nat64_create(), so they are set to the full range of 1 to 65535 for all
the variants of configuring the overload addresses.

VRVDR-50828