nat64: make per-rule 'used' count be decremented
The 'used' count displayed by 'show nat nat64 rules' for overload
configuration kept going up, as it was not being decremented. This was
because the rule was not passed into npf_nat_free_map() when releasing a
port map entry.  This is fixed here by passing the rule into the function.

VRVDR-50828