vlan_modify: fix json writing unwinding for no interfaces
Avoid going into a large, but finite loop, when there are no
interfaces by ending more depths than were opened, causing the depth
field to underflow and go to UINT32_MAX and subsequent indentation to
take a long time and memory.

The starting of an object and an array is only done if there is at
least one match in the filter chain head list iteration, so only undo
this if at least one interface was found.

Fixes: VRVDR-50839