route: Add an api to allow plugins to provide path state
Add API to allow plugins or internal code to notify that a path
has become unusable. Add API to allow the core code to query the
usability of a path.

An interface may go down in which case the core code needs to know
about that asap so that it can cut over from primary paths to backup
paths if the primary is no longer usable. Features that track state
can call this new API to inform the core code of this as soon as it
happens.  Features that use this api to inform the core code that a
path is down should also register a function so that they can be
queried about the state of a path.  If the feature does not know the
state it should return 'unknown'

Add the implementation to a new file ip_route as this is for v4 and v6.
This will become the file that the common v4/v6 route code lives in.

VRVDR-50721