nh_common: use a nh_map when a next_hop_list has backup paths
Add a new object, a next hop map, that can be used to help selecting
a path from a list of paths. It provides the following benefits:
- Allows a quick atomic switchover from using a path to not using it.
  This is needed when a primary path becomes unusable and we have to
  switch over to the backup
- Allows fairness across paths when a primary path becomes unuseable
  and there are further primaries available.

The map contains an array of offsets. The offset is to the netxthop
to use. For example if we have a next_hop_list with 4 primary paths and 1
backup we could have:

 nh0, nh2, nh3 and primary.  nh1 backup.
 map: [0] val 0
      [1] val 2
      [2] val 3
      [3] val 4

We then hash over the number of entries in the map, and get the value
for that index.  In the case above the hash value of 0 would mean use
nh0. A hash value of 1 would mean use nh2.

To give us fairness after a cutover we want to have each remaining
active primary to get an equal share of the traffic, or each backup to
get an equal share.

If for every primary path we start with (#primaries -1) entries in
the map for that primary, then when a primary becomes unusable we
can allocate all the entries for that primary equally to the other
remaining primaries.  To do this we allocate (#primary *
(#primary -1)) paths in the map, with a max of 64 so that it still
fits in a single cache line.

VRVDR-50722