ip_rt_protobuf: preserve display behaviour for MPLS deagg routes
Previously the RIB would generate a route update for a deagg route
with an RTA_VIA attribute present that would then cause the dataplane
to add a nexthop with the RTF_GATEWAY flag set, and this would then
result in the following label-table output to be displayed:

    in label: 53764, fec:ipv6
            nexthop via ::, vrfred

With protobufs, this has been fixed such that no nexthop field is
present in the protobuf case, but this results in a change in
output. Therefore, preserve that by setting the RTF_GATEWAY flag for
deagg routes, ensuring that the nexthop address family also is
modified to that expected. Whether this is the best way to display a
deagg route can then be followed up separately.

VRVDR-50720