nh_common: change the nexthop to have a struct_ip addr
Previously the next_hop had a union of address types, but did not
have the field to indicate if the address was of type v4 or v6.

Change it to have an ip_addr so that the type can be stored there.
This changes the size of the strure and makes the next_hop_list
spill into a 3rd cache line, but the important fields for performance
do not move out of the first cache line.

VRVDR-50724