nh_common: provide apis to help do a modify of an active next_hop_list
There is a need to replace a next_hop_list that is in use in the
forwarding path when the l2 neigh for the nexthop is resolved/unresolved.
Currently this is done by v4 and v6 route code, and they need to know too
much about the implementation of the next_hops.  Move the code that does
this into new functions and have the v4/v6 code call these new APIs to
do the modify/replace.

VRVDR-50724