nh_common: store the protected next_hops in a 2 level hash
Store the protected next_hops in a 2 level hash to allow for quick
lookup when a link or link/nh pair becomes unusable.

The first level hash is for the interface, where the key is the
ifindex.  Each entry in the interface hash has a 2nd hash table,
where the key is the gateway, or 0 if there is no gateway. The
result of a lookup in the 2nd hash is a structure that has a list of
all the nexthops that use that interface/gateway pair.

When an interface moves to 'down' the interface can be looked up in
the first level hash table, and all next_hops using that interface
can be found by walking over the hash and then walking over the list
for each entry.

In the case where a interface/next_hop pair is marked unusable then
the list of next hops using that is found by doing 2 hash lookups.

VRVDR-50722