debug_strip: add build-ids to package metadata
Collect the build-ids for a package and put them in
debian/.debhelper/$package/dbgsym-build-ids. This allows them to be
picked up by dh_gencontrol and added to the package metadata, in the
same way as for auto-generated dbgsym packages, and are useful in
verifying that the binary and -dbg packages match.