nh_common: update sw forwarding state when a NH becomes unusable
When a next_hop is marked as unusable via the new API, modify the
forwarding state. The 2 level hash lookup gives a list of all the
NHs that need to be modifed. Walk over all the entries to get the
next_hops that are now unusable.

From the next_hop it walks back to the next_hop_list to modify the
nh_map. When modifying the map, if there are still usable primary
paths then use them, otherwise move them all to the backup path(s).

Now that the nh_map can be updated from a different thread we need
to recheck the usability at the end of creating new next_hop_lists.
This is done at the end of adding the tracking in the initial
create case, and at the end of finishig the copy in the case of
a modify of one that was already in the forwarding path.

VRVDR-50722