nh_common: add a ptr back from the NH to the NH list
When a next hop becomes unusable the lookup will take us to a NH,
but we need to get back to the next_hop_list to do the required
updates. Add a pointer from the next_hop back to the next_hop_list
so that the updates can be done properly.

VRVDR-50722