NPF: Use ICMP opcode defines, not magic values
Jira: VRVDR-50158

We had previously defined names for the bits in the ncode operand
which indicated if a type and/or code was being matched, and to
set or extract the values.

Use those when executing the ICMP match opcodes.
Also fix the bracing style in that match code.

No functional change.