if: make promiscuity apply to VLANs as well as MAC addresses
Some DPDK drivers (e.g. vmxnet3) don't cope well with disabling VLAN
promiscuous mode when in MAC promiscuous mode. We also don't have a
case where we want one instead of both, since when there is no support
for programming VLAN filtering of DPDK PMD ports when using a switch
interface.

So define interface promiscuity in the dataplane as applying to VLANs
as well as MAC addresses, as this matches Linux kernel semantics,
there is little benefit in separating the two levels of filtering, and
it works around the DPDK driver issues, and remove the
enabling/disabling of VLAN filtering during capture that is now
redundant.

Fixes: VRVDR-52049