urcu: add an API to allow a thread to register with urcu
With pic_edge code there will be a requirement to access 'state' from
threads such as platform interrupt threads to meet the cutover time
requirements. As these threads will be accessing rcu controlled state
they must be registered with rcu. There is no rcu api to check if
a thread is already registered, so provide a new api that uses
thread local storage to track whether it is already registered and
brings the thread online if it is.

VRVDR-50722