bridge: issuing FAL delport notifications for members on bridge delete
Currently, if a bridge is deleted when there are active members (which
can happen during delete or other interface cleanup), then the memory
for the bridge port is freed and the brport field in the member ifp
NULLd, but other cleanup isn't done, such as FAL notifications and
resetting promiscuity.

Issuing FAL notifications in bridge_if_uninit is too late because the
FAL L2 delete has already been issued for the bridge/switch and there
could still be references to it. Therefore, instead listen to the
L2_DELETED event and clean up the members of the bridge at that point,
and do all of the cleanup that would be done for a regular removal of
the port from the bridge, including issuing of FAL notifications.

In addition, also do clean up of bridge members that are being
deleted, such that FAL notifications are issued and the brport memory
is freed.

Fixes: VRVDR-51123