npf: Add a flag to enable hash table linkage for rules
NPF rules are currently maintained in an ordered linked list.
The ordering is critical both for insertion of rules into
grouper as well as packet processing for cases not handled by
grouper. However, it does not scale for lookups in large rulesets.
A parallel linkage mechanism using a hash table needs to be
added to support faster lookups.

Add a flag to enable hash table linkage and enable the flag just
for IPsec rules.

VRVDR-50304