nh_common: add a bitmask to track usable next_hops
When updating the next hop map based on the given index being
unusable walk across all entries, and if the map contains the
index then replace it with the next value.  The next value is
based on around robin over the remaining usable primary paths.

As there may be multiple threads updating this map at the same
time we need to make sure that we end up in the correct final
state. For example, if we started with: 0,1,2,3,0,1,2,3,0,1,2,3
and then are removing 0 and 1 at the same times (from different
threads) then we need to be sure that at the end of the pass
removing index 0 that we haven't put references to index1 in
that were not caught by the other thread as it was slightly
ahead.

To do this we maintains an atomic bitmask of the usable paths,
and verify at the end of the loop that only the index we are
processing has been unset.  If another index has become unusable
then we need to go round the loop again, but this time removing
references to the complete set of unusable paths, and choosing
the new value from the remaining usable paths.

VRVDR-50722