nh_common: reinit nh map contents if collisions when marking unusable
If multiple threads are updating the map to mark it unusable then
it can get into an unknown state and attempts to correct it when looping
round again on losing the race may leave the map in an unfair state.

For example, if we have an initial map of 0,1,2,3,0,1,2,3,0,1,2,3
and we have 2 threads, removing 0 and 1 then we can get to the stage
where after we have removed 1 we get: 0,0,2,3,0,2,2,3,0,3,2,3 but
that could be followed immediately by removal of 0 (at which stage
processing of 1 hasn't finished so 1 is still seen as usable) which
gives: 1,2,2,3,3,2,2,3,1,3,2,3. With this the thread that loses the
would go round a second time, and remove the set of the values that
were being removed, giving: 2,2,2,3,3,2,2,3,3,3,2,3 which is a fair
result.

However, if we were also having to factor in adding of paths as usable
then we would want to make sure that we were back in a known state
when we added and removed paths at the same time. To make sure that
we end up in a fair state, if there is a race then rebuild the map
as if it was the initila build.

VRVDR-50722