nh_common: let paths be marked usable and unusable
Change the code to allow a next_hop to be marked usable as well as
unusable.  This is mostly passing an extra value through the
functions until we get to the code to rebuild the nh map.
There we add a new function to update the map if it is an add. This
function works out how many paths it needs to set and walks through
the existing map, setting it fairly. On a collision between
threads the loser of the race will loop again setting the contents
as if it was the initial init.

VRVDR-50722