nat64: check rule group name of rule is set before accessing it
The group of rules associated with the a NAT64 rule can be NULL when
reconfiguration occurs, as the rule is disassociated from the group. This
means that it cannot find the rule group name, and so npf_rule_get_name()
returns NULL.  Therefore, before attempting to print the rule group
returned by npf_rule_get_name() in JSON output, it needs to ensure it
is not NULL. If it is NULL, it will use name "<UNKNOWN>".

VRVDR-51170