nd6_nbr: when storing the v6 addr use sockaddr_storage not sockaddr
The ns input code uses a struct sockaddr on the stack to store the
v6 address of the packet, and then passes the address of that sockaddr
to another function for it to verify the address.  This is writing
over the end of the sockaddr memory. It should use a sockaddr_storage
instead.

VRVDR-50825