NPF: Enable matching of ICMP classes in ncode
Jira: VRVDR-50158

Add the infrastructure to generate and execute a variant upon
the ICMP ncode instruction such that we can match upon the
'class' of an IPv4 ICMP or ICMPv6 packet.

Where 'class' is either 'error' or 'info'.

Currently we disable grouper matching for ICMP type/code if
ICMP class matching is selected.  It will eventually be possible
to enable grouper support for class matching of ICMPv6, as the
'type' range assignment is suitable; this will not be possible
for IPv4 ICMP.

Some more work is necessary in the cache building before this
is usable for IPv4. That is because 'redirect' is incorrectly
marked as being an error, and we should possibly include the
obsolete 'source quench' which is currently ignored.

These seem to be related to the logic for NAT translating embedded
ICMP error packets, but for the redirect case this is pointless.