main: add a wrapper to unregister a thread with rcu
The dataplane already has a wrapper that registers a thread with
rcu if it is not already registered. It needs a similar wrapper
for unregistering for the case where threads register/unregister
multiple times.

As there can be multiple calls to this it needs to change to a refcount.

VRVDR-51242