vhost: do not wait forever for QMP
If a guest is shutting down, we might race against the disappearing
vhost interfaces. Regardless, we can not wait forever for a response
from QMP because we are holding the RCU lock. If QMP is unresponsive,
assume we can not make the update we want, note the error and return.

Bugfix: VRVDR-51099