ixgbe: do not use the LSC interrupt
The ixgbe LSC interrupt is somewhat unusual. The interrupt tells
the device driver that the link status is changing. The current
DPDK driver waits 1s on the down -> up transition and checks the
link status. However, other code in the driver implies that it
could take up to 9s for the driver to successfully negotiate the
link status.

Empircal testing has shown that waiting 1s is too short. Instead
of modifying the driver, avoid using the LSC interrupt and poll
periodically.

Bugfix: VRVDR-48438