storm_ctl: make threshold removal FAL state symmetric
Currently, on applying a profile with a subset of traffic type
thresholds set, then only policers for that subset are created and set
on the interface or (interface, VLAN) via FAL updates. However, on
deleting a traffic type from a profile that is set on an interface
then the corresponding policer isn't deleted, meaning that the
sequence of configuration operations with their inverses can lead to a
different final state, compared to the sequence and their inverses not
being performed.

Whilst this could be rectified by creating the additional policer
objects in advance and setting their FAL_POLICER_ATTR_CIR attributes
to 0, the use of 0 to mean infinite isn't well-defined in the API, and
this results in more resources being used (i.e. policer objects), as
well as being less flexible in case there is a platform in the future
that only supports a subset (e.g. unicast & multicast).

Therefore, rectify the asymmetry of the operations by unapplying the
policer object for the traffic type upon modifying a profile to remove
a threshold for a traffic type.

Fixes: VRVDR-51406