lpm6: fix check for depth when removing /24
There was an issue in the lpm6 code when deleting a /24 as the check
was for a depth < 24, and the else case was for when the depth was
greater 24.

This lead to the behaviour of deleting the rule from the lpm, but
leaving the /24 a a valid entry in the lpm, so that when a lookup
was done it could return a next_hop index indicating that the lookup
was successful, even though that next_hop index could have been set
to NULL if the /24 was the last user of that next_hop. This could
lead to a crash if the lookup lead to dereferencing a NULL next_hop
as that is never expected to be the case.

Added unit test that mirrored the set of updates that were seen in
the case where the crash happened on a live system.

VRVDR-51385